requirements.txt
flask==3.0.0
flask-cors==4.0.0
spacy==3.7.2
nltk==3.8.1
requests==2.31.0
textblob==0.17.1
python-dotenv==1.0.0
